rag_llm_service/requirements.txt
# Lista zależności dla mikroserwisu LLM

fastapi
uvicorn
openai
python-dotenv
pytest



#TODO: change template